script/requirements.txt
gspread
requests
oauth2client
pandas
python-dateutil
python-dotenv
Telethon
cachetools